[#131] Refactor `Domain` and `Data` module
