+++ CoroutineTemplate/app/src/main/java/co/nimblehq/coroutine/di/modules/DispatcherModule.kt
@@ -1,28 +1,18 @@
 package co.nimblehq.coroutine.di.modules
 
-import co.nimblehq.coroutine.domain.dispatcher.DefaultDispatcher
-import co.nimblehq.coroutine.domain.dispatcher.IoDispatcher
-import co.nimblehq.coroutine.domain.dispatcher.MainDispatcher
+import co.nimblehq.coroutine.dispatcher.DispatchersProvider
+import co.nimblehq.coroutine.dispatcher.DispatchersProviderImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
 
 @Module
 @InstallIn(SingletonComponent::class)
 class DispatcherModule {
 
-    @DefaultDispatcher
     @Provides
-    fun provideDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
-
-    @IoDispatcher
-    @Provides
-    fun provideIoDispatcher(): CoroutineDispatcher = Dispatchers.IO
-
-    @MainDispatcher
-    @Provides
-    fun provideMainDispatcher(): CoroutineDispatcher = Dispatchers.Main
+    fun provideDispatchersProvider(): DispatchersProvider {
+        return DispatchersProviderImpl()
+    }
 }
