+++ CoroutineTemplate/app/src/main/java/co/nimblehq/coroutine/dispatcher/DispatchersProvider.kt
@@ -1,0 +1,9 @@
+package co.nimblehq.coroutine.dispatcher
+
+import kotlinx.coroutines.CoroutineDispatcher
+
+interface DispatchersProvider {
+    fun getIO(): CoroutineDispatcher
+    fun getMain(): CoroutineDispatcher
+    fun getDefault(): CoroutineDispatcher
+}
