+++ CoroutineTemplate/app/src/main/java/co/nimblehq/coroutine/ui/screens/MainViewModel.kt
@@ -1,8 +1,10 @@
 package co.nimblehq.coroutine.ui.screens
 
+import co.nimblehq.coroutine.dispatcher.DispatchersProvider
 import co.nimblehq.coroutine.ui.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class MainViewModel @Inject constructor() : BaseViewModel()
+class MainViewModel @Inject constructor(dispatchersProvider: DispatchersProvider) :
+    BaseViewModel(dispatchersProvider)
