+++ CoroutineTemplate/app/src/main/java/co/nimblehq/coroutine/ui/screens/compose/composables/ComposeScreen.kt
@@ -7,8 +7,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import co.nimblehq.coroutine.domain.data.entity.UserEntity
 import co.nimblehq.coroutine.ui.screens.compose.theme.ComposeTheme
+import co.nimblehq.domain.entity.UserEntity
 
 @Suppress("LongMethod", "LongParameterList")
 @ExperimentalComposeUiApi
