+++ CoroutineTemplate/app/src/main/java/co/nimblehq/coroutine/ui/screens/home/HomeViewModel.kt
@@ -1,12 +1,12 @@
 package co.nimblehq.coroutine.ui.screens.home
 
 import androidx.lifecycle.viewModelScope
-import co.nimblehq.coroutine.domain.data.entity.UserEntity
-import co.nimblehq.coroutine.domain.usecase.GetUsersUseCase
-import co.nimblehq.coroutine.domain.usecase.UseCaseResult
+import co.nimblehq.coroutine.dispatcher.DispatchersProvider
 import co.nimblehq.coroutine.ui.base.BaseViewModel
 import co.nimblehq.coroutine.ui.base.NavigationEvent
 import co.nimblehq.coroutine.ui.screens.second.SecondBundle
+import co.nimblehq.domain.entity.UserEntity
+import co.nimblehq.domain.usecase.GetUsersUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -24,8 +24,9 @@ interface Output {
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
-    private val getUsersUseCase: GetUsersUseCase
-) : BaseViewModel(), Output {
+    private val getUsersUseCase: GetUsersUseCase,
+    dispatchers: DispatchersProvider
+) : BaseViewModel(dispatchers), Output {
 
     private val _users = MutableStateFlow<List<UserEntity>>(emptyList())
     override val users: StateFlow<List<UserEntity>>
@@ -49,10 +50,10 @@ class HomeViewModel @Inject constructor(
 
     private fun fetchUsers() {
         showLoading()
-        viewModelScope.launch {
-            when (val result = getUsersUseCase.execute()) {
-                is UseCaseResult.Success -> _users.value = result.data
-                is UseCaseResult.Error -> _error.emit(result.exception.message.orEmpty())
+        execute {
+            when (val result = getUsersUseCase.execute(page = 1, size = 10)) {
+                is co.nimblehq.domain.usecase.UseCaseResult.Success -> _users.value = result.data
+                is co.nimblehq.domain.usecase.UseCaseResult.Error -> _error.emit(result.exception.message.orEmpty())
             }
             hideLoading()
         }
