+++ CoroutineTemplate/app/src/main/java/co/nimblehq/coroutine/ui/screens/second/SecondViewModel.kt
@@ -1,5 +1,9 @@
 package co.nimblehq.coroutine.ui.screens.second
 
+import co.nimblehq.coroutine.dispatcher.DispatchersProvider
 import co.nimblehq.coroutine.ui.base.BaseViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
 
-class SecondViewModel : BaseViewModel()
+@HiltViewModel
+class SecondViewModel @Inject constructor(dispatchers: DispatchersProvider) : BaseViewModel(dispatchers)
