+++ CoroutineTemplate/app/src/main/res/drawable-v24/ic_launcher_foreground.xml
@@ -1,9 +1,9 @@
 <vector xmlns:aapt="http://schemas.android.com/aapt"
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:width="108dp"
-        android:height="108dp"
-        android:viewportHeight="108"
-        android:viewportWidth="108">
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="108dp"
+    android:height="108dp"
+    android:viewportHeight="108"
+    android:viewportWidth="108">
     <path
         android:fillType="evenOdd"
         android:pathData="M32,64C32,64 38.39,52.99 44.13,50.95C51.37,48.37 70.14,49.57 70.14,49.57L108.26,87.69L108,109.01L75.97,107.97L32,64Z"
@@ -18,10 +18,10 @@
                 android:type="linear">
                 <item
                     android:color="#44000000"
-                    android:offset="0.0"/>
+                    android:offset="0.0" />
                 <item
                     android:color="#00000000"
-                    android:offset="1.0"/>
+                    android:offset="1.0" />
             </gradient>
         </aapt:attr>
     </path>
@@ -30,5 +30,5 @@
         android:fillType="nonZero"
         android:pathData="M66.94,46.02L66.94,46.02C72.44,50.07 76,56.61 76,64L32,64C32,56.61 35.56,50.11 40.98,46.06L36.18,41.19C35.45,40.45 35.45,39.3 36.18,38.56C36.91,37.81 38.05,37.81 38.78,38.56L44.25,44.05C47.18,42.57 50.48,41.71 54,41.71C57.48,41.71 60.78,42.57 63.68,44.05L69.11,38.56C69.84,37.81 70.98,37.81 71.71,38.56C72.44,39.3 72.44,40.45 71.71,41.19L66.94,46.02ZM62.94,56.92C64.08,56.92 65,56.01 65,54.88C65,53.76 64.08,52.85 62.94,52.85C61.8,52.85 60.88,53.76 60.88,54.88C60.88,56.01 61.8,56.92 62.94,56.92ZM45.06,56.92C46.2,56.92 47.13,56.01 47.13,54.88C47.13,53.76 46.2,52.85 45.06,52.85C43.92,52.85 43,53.76 43,54.88C43,56.01 43.92,56.92 45.06,56.92Z"
         android:strokeColor="#00000000"
-        android:strokeWidth="1"/>
+        android:strokeWidth="1" />
 </vector>
