+++ CoroutineTemplate/build.gradle
@@ -19,12 +19,12 @@ buildscript {
         androidx_navigation_version = '2.3.4'
         androidx_support_version = '1.3.0'
 
-        compose_version = '1.0.0-rc01'
+        compose_version = '1.0.2'
         constraint_layout_version = '2.0.0-alpha3'
 
-        hilt_version = '2.37'
+        hilt_version = '2.38.1'
 
-        kotlin_version = '1.5.10'
+        kotlin_version = '1.5.21'
         kotlinx_coroutines_version = '1.5.0'
 
         moshi_version = '1.12.0'
@@ -32,6 +32,7 @@ buildscript {
         okhttp_version = '4.9.1'
         retrofit_version = '2.9.0'
 
+        mockito_version = '3.3.1'
         test_android_junit_version = '1.1.1'
         test_espresso_version = '3.1.0'
         test_junit_version = '4.13.2'
@@ -40,7 +41,7 @@ buildscript {
 
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
     dependencies {
         classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$androidx_navigation_version"
@@ -57,7 +58,6 @@ plugins {
 allprojects {
     repositories {
         google()
-        jcenter()
         maven { url 'https://jitpack.io' }
         mavenCentral()
     }
