+++ CoroutineTemplate/data/src/main/java/co/nimblehq/coroutine/storage/BaseSharedPreferences.kt
@@ -1,7 +1,6 @@
-package co.nimblehq.coroutine.domain.storage
+package co.nimblehq.coroutine.storage
 
 import android.content.SharedPreferences
-import co.nimblehq.coroutine.domain.extension.execute
 
 abstract class BaseSharedPreferences {
