+++ CoroutineTemplate/data/src/main/java/co/nimblehq/coroutine/storage/NormalSharedPreferences.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.coroutine.domain.storage
+package co.nimblehq.coroutine.storage
 
 import android.content.Context
 import android.content.SharedPreferences
