+++ CoroutineTemplate/data/src/main/java/co/nimblehq/coroutine/storage/SharedPreferencesExt.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.coroutine.domain.extension
+package co.nimblehq.coroutine.storage
 
 import android.content.SharedPreferences
