--- CoroutineTemplate/domain/src/main/java/co/nimblehq/coroutine/domain/dispatcher/DispatcherQualifiers.kt
@@ -1,15 +1,0 @@
-package co.nimblehq.coroutine.domain.dispatcher
-
-import javax.inject.Qualifier
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
-annotation class DefaultDispatcher
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
-annotation class IoDispatcher
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
-annotation class MainDispatcher
