--- CoroutineTemplate/domain/src/main/java/co/nimblehq/coroutine/domain/repository/UserRepository.kt
@@ -1,21 +1,0 @@
-package co.nimblehq.coroutine.domain.repository
-
-import co.nimblehq.coroutine.data.service.ApiService
-import co.nimblehq.coroutine.domain.data.entity.UserEntity
-import co.nimblehq.coroutine.domain.data.entity.toUserEntities
-import javax.inject.Inject
-
-interface UserRepository {
-
-    suspend fun getUsers(): List<UserEntity>
-}
-
-class UserRepositoryImpl @Inject constructor(
-    private val apiService: ApiService
-) : UserRepository {
-
-    override suspend fun getUsers(): List<UserEntity> {
-        val response = apiService.getUsers()
-        return response.toUserEntities()
-    }
-}
