+++ CoroutineTemplate/domain/src/main/java/co/nimblehq/domain/entity/UserEntity.kt
@@ -1,0 +1,26 @@
+package co.nimblehq.domain.entity
+
+data class UserEntity(
+    val id: Int?,
+    val name: String,
+    val username: String,
+    val email: String?,
+    val address: Address?,
+    val phone: String,
+    val website: String
+) {
+
+    data class Address(
+        val street: String,
+        val suite: String,
+        val city: String,
+        val zipCode: String,
+        val geo: Geo?
+    ) {
+
+        data class Geo(
+            val latitude: String,
+            val longitude: String
+        )
+    }
+}
