+++ CoroutineTemplate/gradle.properties
@@ -13,3 +13,5 @@ org.gradle.jvmargs=-Xmx1536m
 # org.gradle.parallel=true
 android.enableJetifier=true
 android.useAndroidX=true
+#https://kotlinlang.org/docs/kapt.html#running-kapt-tasks-in-parallel
+kapt.use.worker.api=true
